requirements.txt
affine==2.4.0
annotated-types==0.7.0
asttokens==3.0.0
attrs==25.1.0
blis==1.3.3
cartopy==0.24.1
catalogue==2.0.10
certifi==2025.11.12
cffi==1.17.1
charset-normalizer==3.4.4
click==8.3.1
click-plugins==1.1.1
cligj==0.7.2
cloudpathlib==0.23.0
comm==0.2.2
confection==0.1.5
contextily==1.6.2
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
cymem==2.0.13
cython==0.29.37
debugpy==1.8.13
decorator==5.2.1
dol==0.3.14
dotenv==0.9.9
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
executing==2.2.0
fonttools==4.56.0
gcloud==0.18.3
geographiclib==2.0
geopandas==1.0.1
geoplot==0.5.1
geopy==2.4.1
googleapis-common-protos==1.69.2
hf==0.0.8
httplib2==0.22.0
idna==3.11
ipykernel==6.29.5
ipython==9.0.2
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
joblib==1.4.2
jupyter-client==8.6.3
jupyter-core==5.7.2
jwcrypto==1.5.6
kiwisolver==1.4.8
logger==1.4
mapclassify==2.8.1
markupsafe==3.0.3
matplotlib==3.10.1
matplotlib-inline==0.1.7
mercantile==1.2.1
murmurhash==1.0.15
narwhals==1.30.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.4.0
oauth2client==4.1.3
oauthlib==3.2.2
packaging==25.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
pip==25.3
platformdirs==4.3.6
plotly==6.0.0
praw==7.8.1
prawcore==2.4.0
preshed==3.0.12
prompt-toolkit==3.0.50
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyahocorasick==2.1.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.12.5
pydantic-core==2.41.5
pygments==2.19.1
pyogrio==0.10.0
pyparsing==3.2.1
pyproj==3.7.1
pyrebase4==4.8.0
pyshp==2.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jwt==4.1.0
pytz==2025.1
pyzmq==26.2.1
rasterio==1.4.3
ratelimiter==1.2.0.post0
regex==2024.11.6
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==0.10.1
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==80.9.0
shapely==2.0.7
six==1.17.0
smart-open==7.5.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
stack-data==0.6.3
thinc==8.3.10
threadpoolctl==3.5.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
tweepy==4.15.0
typer-slim==0.21.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.1
update-checker==0.18.0
urllib3==2.6.2
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.3
websocket-client==1.8.0
wheel==0.45.1
wrapt==2.0.1
xyzservices==2025.1.0
